requirements.txt
# Core PDF Processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
camelot-py[cv]>=0.11.0

# OCR & Image Processing
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0

# AI APIs
anthropic>=0.25.0
openai>=1.0.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
arrow>=1.3.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# CLI & UI
rich>=13.0.0
click>=8.1.0
streamlit>=1.28.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
ipython>=8.15.0
ipdb>=0.13.13

# Security
pip-audit>=2.6.0
